+++ devtools/create_titanic/base-str.cpp
@@ -1,0 +1,2 @@
+#define SCUMMVM_UTIL 1
+#include "../common/base-str.cpp"
