+++ devtools/create_xeen/base-str.cpp
@@ -1,0 +1,2 @@
+#define SCUMMVM_UTIL 1
+#include "../common/base-str.cpp"
