+++ engines/glk/advsys/glk_interface.cpp
@@ -41,7 +41,7 @@ void GlkInterface::print(const Common::U32String &msg) {
 	// Don't print out text if loading a savegame directly from the launcher, since we don't
 	// want any of the intro text displayed by the startup code to show
 	if (_saveSlot == -1)
-		glk_put_string_stream_uni(glk_window_get_stream(_window), msg.c_str());
+		glk_put_string_stream_uni(glk_window_get_stream(_window), msg.u32_str());
 }
 
 void GlkInterface::print(int number) {
